+++ src/main/java/com/kang/blog/repository/BoardRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import javax.persistence.LockModeType;
 import java.util.Optional;
 
-public interface BoardRepository extends JpaRepository<Board, Integer>, BoardRepositoryCustom {
+public interface BoardRepository extends JpaRepository<Board, Long>, BoardRepositoryCustom {
 
       //조회수 증가 - 동시성 문제 해결법- Udate 쿼리 방법
 //    @Modifying   // Read가 아닌 C,U,D의 경우 @Modifying이 필요.
@@ -20,5 +20,5 @@ public interface BoardRepository extends JpaRepository<Board, Integer>, BoardRep
     //조회수 증가 - 동시성 문제 해결법- 비관적 락 방법
     @Lock(LockModeType.PESSIMISTIC_WRITE)
     @Query(value = "SELECT b FROM Board b WHERE b.id = :id")
-    Optional<Board> findByIdForUpdate(@Param("id") int id);
+    Optional<Board> findByIdForUpdate(@Param("id") Long id);
 }
